+++ java-baseball/src/main/java/baseball/domain/NumbersGenerator.java
@@ -1,8 +1,19 @@
 package baseball.domain;
 
+import camp.nextstep.edu.missionutils.Randoms;
+import java.util.ArrayList;
 import java.util.List;
 
-public interface NumbersGenerator {
+public class NumbersGenerator {
 
-    List<Integer> makeComputer();
+    public List<Integer> makeComputer() {
+        List<Integer> computer = new ArrayList<>();
+        while (computer.size() < 3) {
+            int randomNumber = Randoms.pickNumberInRange(1, 9);
+            if (!computer.contains(randomNumber)) {
+                computer.add(randomNumber);
+            }
+        }
+        return computer;
+    }
 }
