--- java-baseball/src/main/java/baseball/domain/RandomNumbersGenerator.java
@@ -1,20 +1,0 @@
-package baseball.domain;
-
-import camp.nextstep.edu.missionutils.Randoms;
-import java.util.ArrayList;
-import java.util.List;
-
-public class RandomNumbersGenerator implements NumbersGenerator {
-
-    @Override
-    public List<Integer> makeComputer() {
-        List<Integer> computer = new ArrayList<>();
-        while (computer.size() < 3) {
-            int randomNumber = Randoms.pickNumberInRange(1, 9);
-            if (!computer.contains(randomNumber)) {
-                computer.add(randomNumber);
-            }
-        }
-        return computer;
-    }
-}
